Revert "Complete port to Doubles instead of Floats"
This reverts commit b863315f4a202ad44e455ab708b43a298983cf3a.

+++ src/Jitter2/Collision/CollisionFilter/INarrowPhaseFilter.cs
@@ -39,5 +39,5 @@ public interface INarrowPhaseFilter
     /// <returns>False if the collision should be filtered out, true otherwise.</returns>
     bool Filter(RigidBodyShape shapeA, RigidBodyShape shapeB,
         ref JVector pointA, ref JVector pointB,
-        ref JVector normal, ref double penetration);
+        ref JVector normal, ref float penetration);
 }
