+++ src/Jitter2/SoftBodies/BroadPhaseCollisionFilter.cs
@@ -52,7 +52,7 @@ public class BroadPhaseCollisionFilter : IBroadPhaseFilter
 
             bool colliding = NarrowPhase.MPREPA(i1, i2,
                 JQuaternion.Identity, JVector.Zero,
-                out JVector pA, out JVector pB, out JVector normal, out double penetration);
+                out JVector pA, out JVector pB, out JVector normal, out float penetration);
 
             if (!colliding) return false;
 
@@ -72,7 +72,7 @@ public class BroadPhaseCollisionFilter : IBroadPhaseFilter
             if (!i1.SoftBody.IsActive && !rb.Data.IsActive) return false;
 
             bool colliding = NarrowPhase.MPREPA(i1, (proxyB as RigidBodyShape)!, rb.Orientation, rb.Position,
-                out JVector pA, out JVector pB, out JVector normal, out double penetration);
+                out JVector pA, out JVector pB, out JVector normal, out float penetration);
 
             if (!colliding) return false;
 
@@ -91,7 +91,7 @@ public class BroadPhaseCollisionFilter : IBroadPhaseFilter
             if (!i2.SoftBody.IsActive && !ra.Data.IsActive) return false;
 
             bool colliding = NarrowPhase.MPREPA(i2, (proxyA as RigidBodyShape)!, ra.Orientation, ra.Position,
-                out JVector pA, out JVector pB, out JVector normal, out double penetration);
+                out JVector pA, out JVector pB, out JVector normal, out float penetration);
 
             if (!colliding) return false;
